+++ pages/api/sites/index.ts
@@ -1,17 +1,16 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import { assertAuth } from "../../../src/server/lib/api-helpers";
-import { getSites } from "../../../src/server/lib/site";
-import {CACHE_VERY_SHORT} from "../../../config.web";
+import { NextApiRequest, NextApiResponse } from 'next'
+import { assertAuth } from '../../../src/server/lib/api-helpers'
+import { getSites } from '../../../src/server/lib/site'
 
 export const ProjectsHandler = async (request: NextApiRequest, response: NextApiResponse) => {
     if (request.method === 'GET') {
-        const sites = await getSites();
+        const sites = await getSites(request.query.active_only === 'true')
         response.setHeader(
           'Cache-Control',
           'public, s-maxage=30, stale-while-revalidate=59'
-        );
-        response.send(sites);
-        return;
+        )
+        response.send(sites)
+        return
     }
 }